move(schedule): organization 도메인에서 사용하는 DTO가 schedule에 정의되어 있어 수정
