+++ src/main/java/com/umc/product/schedule/adapter/in/web/v1/ScheduleCommandV1CommandV1Controller.java
@@ -6,9 +6,9 @@ import com.umc.product.authorization.domain.ResourceType;
 import com.umc.product.global.exception.NotImplementedException;
 import com.umc.product.global.security.MemberPrincipal;
 import com.umc.product.global.security.annotation.CurrentMember;
+import com.umc.product.organization.adapter.in.web.dto.request.CreateStudyGroupScheduleRequest;
 import com.umc.product.schedule.adapter.in.web.v1.dto.request.CreateScheduleRequest;
 import com.umc.product.schedule.adapter.in.web.v1.dto.request.CreateScheduleWithAttendanceRequest;
-import com.umc.product.schedule.adapter.in.web.v1.dto.request.CreateStudyGroupScheduleRequest;
 import com.umc.product.schedule.adapter.in.web.v1.dto.request.UpdateScheduleLocationRequest;
 import com.umc.product.schedule.adapter.in.web.v1.dto.request.UpdateScheduleRequest;
 import com.umc.product.schedule.adapter.in.web.v1.dto.response.UpdateScheduleLocationResponse;
